refactor: separate UI helpers
